move read_ecsv_keys() from fba_launch_io.py to utils.py
